Modified project structure so that it could become more clear
