+++ MoviesApp/MoviesApp.xcworkspace/xcuserdata/andrefaruolo.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -10,7 +10,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "MoviesApp/Model/MovieDAO.swift"
-            timestampString = "564612501.183315"
+            timestampString = "564613211.798565"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "163"
