+++ MoviesApp/MoviesApp/DAOs & Managers/FilterManager.swift
