+++ MoviesApp/MoviesApp/Model/Dates.swift
@@ -1,0 +1,14 @@
+//
+//  Dates.swift
+//  MoviesApp
+//
+//  Created by Andre Faruolo on 22/11/18.
+//  Copyright © 2018 Andre Faruolo. All rights reserved.
+//
+
+import Foundation
+
+struct Dates: DataObject {
+    var maximum: String
+    var minimum: String
+}
