+++ MoviesApp/MoviesApp/Model/Genre.swift
@@ -1,0 +1,16 @@
+//
+//  Genre.swift
+//  MoviesApp
+//
+//  Created by Andre Faruolo on 22/11/18.
+//  Copyright © 2018 Andre Faruolo. All rights reserved.
+//
+
+import Foundation
+
+struct Genre: DataObject{
+    
+    var id: Int
+    var name: String
+    
+}
