+++ MoviesApp/MoviesApp/Model/Genres.swift
@@ -1,0 +1,15 @@
+//
+//  Genres.swift
+//  MoviesApp
+//
+//  Created by Andre Faruolo on 22/11/18.
+//  Copyright © 2018 Andre Faruolo. All rights reserved.
+//
+
+import Foundation
+
+struct Genres: DataObject{
+    
+    var genres: [Genre]
+    
+}
