+++ MoviesApp/MoviesApp/Model/Results.swift
@@ -1,0 +1,13 @@
+//
+//  Results.swift
+//  MoviesApp
+//
+//  Created by Andre Faruolo on 22/11/18.
+//  Copyright © 2018 Andre Faruolo. All rights reserved.
+//
+
+import Foundation
+
+struct Results: DataObject {
+    var movies: [Movie]
+}
