+++ MoviesApp/MoviesApp/Model/Structs.swift
@@ -1,0 +1,24 @@
+//
+//  Movie.swift
+//  MoviesApp
+//
+//  Created by Andre Faruolo on 11/11/18.
+//  Copyright © 2018 Andre Faruolo. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+
+
+
+
+
+
+
+
+
+
+
+
+
